% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_proton.R
\name{add_proton}
\alias{add_proton}
\title{add_proton}
\usage{
add_proton(mass)
}
\arguments{
\item{mass}{a numeric containing a monoisotopic mass}
}
\value{
a numeric containing the monoisotopic mass after proton addition
}
\description{
This function adds the mass of a proton to a monoisotopic mass (e.g., to a deprotonated m/z to obtain its neutral analogue).
}
\examples{
mz <- 412.9664                ## deprotonated PFOA with small error
add_proton(mz)
}
