% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neutral_formulas.R
\name{get_neutral_formulas}
\alias{get_neutral_formulas}
\title{get_neutral_formulas}
\usage{
get_neutral_formulas(
  neutral_mass,
  num_formulas = 10,
  mfRange = "C3-50 H0-100 O0-10 F2-30"
)
}
\arguments{
\item{neutral_mass}{a numeric containing the neutral monoisotopic mass of a chemical compound}

\item{num_formulas}{an integer indicating the maximum number of elemental compositions to return (default = 10)}

\item{mfRange}{a character string defining the range of expected elemental compositions (default = "C3-50 H0-100 O0-10 F2-30" for PFC characterization)}
}
\value{
a tibble of elemental compositions
}
\description{
This function takes a neutral monoisotopic mass and returns a series of elemental compositions, up to the maximum number of formulas specified in the num_formulas parameter.
}
\examples{
neutral_mass <- 413.9735            ## C8HF15O2 (PFOA) with small error
get_neutral_formulas(neutral_mass)
}
