% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neutral_monoisotopic.R
\name{get_neutral_monoisotopic}
\alias{get_neutral_monoisotopic}
\title{get_neutral_monoisotopic}
\usage{
get_neutral_monoisotopic(formula)
}
\arguments{
\item{formula}{a character string containing a molecular formula}
}
\value{
numeric containing the neutral monoisotopic mass
}
\description{
This function converts a molecular formula into its neutral monoisotopic mass.
}
\examples{
formula = "C8 H1 F15 O2"    ## PFOA with spaces
get_neutral_monoisotopic(formula)

formula = "C8HO2F15"        ## PFOA without spaces and subscript = 1
get_neutral_monoisotopic(formula)

formula = "H1 O2 F15 C8"    ## PFOA scrambled
get_neutral_monoisotopic(formula)

formula = "C27H46O"         ## cholesterol
get_neutral_monoisotopic(formula)

formula = "C27H40OD6"       ## d6-cholesterol; returns D as [2H]
get_neutral_monoisotopic(formula)
}
