% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_get_neutral_monoisotopic.R
\name{multiple_get_neutral_monoisotopic}
\alias{multiple_get_neutral_monoisotopic}
\title{multiple_get_neutral_monoisotopic}
\usage{
multiple_get_neutral_monoisotopic(formulas)
}
\arguments{
\item{formulas}{a vector of character strings containing molecular formulas}
}
\value{
a tibble containing the molecular formulas and their monoisotopic masses
}
\description{
This function applies the get_neutral_monoisotopic() function to a group of molecular formulas, rather than a single molecular formula.
}
\examples{
formulas <- c("C8HF15O2",              ##  413.9735 u (PFOA)
              "C9H7F9O2",              ##  413.9735 u
              "C13H7F17O2")            ##  413.9735 u
multiple_get_neutral_monoisotopic(formulas)
}
