% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_proton.R
\name{subtract_proton}
\alias{subtract_proton}
\title{subtract_proton}
\usage{
subtract_proton(mass)
}
\arguments{
\item{mass}{a numeric containing a monoisotopic mass}
}
\value{
a numeric containing the monoisotopic mass after proton subtraction
}
\description{
This function subtracts the mass of a proton from a monoisotopic mass (e.g., from a protonated m/z value to obtain its neutral analogue).
}
\examples{
mz <- 414.9814                ## protonated PFOA with small error
subtract_proton(mz)
}
